Fixed comma operator and function arguments
Fixed by, in the finishCall parsing rule, instead of adding an expression to the arguments, it now adds an equality. This is because an expression has lower precedence, therefore it would be parsing it as a comma. By changing it to equality which has higher precedence than comma, we are making sure that the comma operator is not being parsed inside a function argument list.