Fixed REPL for expressions.
After adding statements, the REPL (command line interpreter) would fail parsing expressions, because it would always try to parse a statement first, as they have higher precedence.

Solution was to add another method to Parser, which will try to parse a list of statements, but is allowed to parse a single expression. If it reaches the end of the sourcee after parsing the expression part of an expression stement, it stops early and returns that expression.

The runPrompt() method was changed to, instead of calling the run() methid with the String with the commands from the command line, it will read the commands line by line from the command line and parse each line individually with the above mentioned method. If it encounters an error, ignore it, because we are in command line mode and don't want to exit on an error (it will still report it) and see if it was returned a list of statements. If so, interpret them as usual. If instead it was returned an expression from the Parser, interpret the expression (with a new method in Interpreter that interprets expressions instead of a List of statements) and print the result, if any.