This completes challenge 3 of chapter 15
The if() in push() affects very negatively the performance. That is why
I am creating this PR instead of merging directlly to master.
Later, I might be able to know how big the stack needs to be before
evaluating the bytecode for a given chunk. This test os best done
at that time.
Until that is possible, this will remain in a separate PR.