Removed optimization from Chapter 21 Challenge 1
This optimization makes it so when an identifier is found, instead
of  creating a new constant (the variable's name), it sees if it
is already in a global variables table. This makes it so we don't
waste memory with new copies of the same global variable in the
constants table.
But this was interfering with the definition of functions.

TODO: it was removed for now, but it should be fixed so that it
doesn't affect declaration of functions.